requirements.txt
numpy>=1.19.0
torch>=1.7.0
torchvision>=0.8.0
matplotlib>=3.3.0
pillow>=8.0.0
scikit-image>=0.17.0
opencv-python>=4.4.0
tqdm>=4.50.0
tensorboard>=2.4.1
albumentations>=0.5.2
PyYAML>=5.4.1
requests>=2.25.1
scipy>=1.6.0
scikit-learn>=0.24.1
